fix(core, api): add response bodies for organization user endpoints - Fixed changeset file - Reorganized tests into two files
